refactor(android): InjectedAgent for the in-app JVMTI agent
MobilecliAgent (entry point) and JsonRpcDispatcher (methods) were one
thing split in two, and the class handling webviews and Flutter was named
after only one of them. Now:

- InjectedAgent: JVMTI entry point plus the JSON-RPC methods it serves
- WebViews: the WebView helpers it calls
- requireParam moves to JsonRpcSocketServer, shared with DeviceServer
